Reenable dynamic linking
This was disabled due to bugs in LLVM 12. The LLVM commit we are using
is much newer and contains the fix (https://reviews.llvm.org/D110261).

The change now lets the host decide whether to link libcxx and
friends statically or dynamically.

Fixes an issue where Gentoo won't be able to build carbon without portage
flags that enable static libraries for libcxx and libunwind.