Detect when evaluating an impl recursively tries to evaluate the same impl for the same or a more complex set of parameters.
This implementation follows the design in #2458.

In order to perform the check after we have tested that the type
structure matches and before we check that constraints are recursively
satisfied, argument deduction is extended to check structural matching
properties earlier.

This requires us to separate match failures into two kinds: hard
failures that produce errors that should never be swallowed, and soft
failures such as a missing impl that lead us to merely discard an impl
as a candidate. A flag has been added to ImplScope and ArgumentDeduction
to specify whether soft failures should produce an error message or not.