Expect () after an alternative only if they were present in its declaration.
The design of choice types expects the declaration of an alternative to
match the usage: if an alternative is declared as `None`, then it should
be used as `None` not `None()`, and if it is declared as `None()` then
it should be used as `None()` not `None`. Update explorer to match.

Also clean up the handling of choice types and alternative values a
little in general, by moving away from identifying choice types and
alternatives as strings and towards identifying them symbolically.

Closes #2422