WIP remove TypeVariableBinding and its subclass MixinSelf
The Self of a Mixin would produce a new type of value which would be a
subclass of VariableType, which would then be checked while interpreting
a simple emember access expression. After the check, the value
corresponding to the mixin Self would be used to search mixin
declarations to find the required member.