Order impl matching by type structure.
As described in
[the generics design](https://github.com/carbon-language/carbon-lang/blob/trunk/docs/design/generics/details.md#type-structure-of-an-impl-declaration)
impl declarations are prioritized by type structure. Given two impls
that match a 'type as interface' query, the one that describes the
longest prefix of the query without using placeholders is preferred.

We implement this by putting all impls in a total order, first by type
structure equivalence classes and then by lexical order. When matching
an impl, we walk this total order, and stop once we find a match and
reach the end of its equivalence class.

Equivalence classes are determined by finding the locations of the
"holes" (the positions where deduced parameters appear) within the type
structure, viewed as a tree. Two impls are in the same equivalence class
if their holes are in the same place, and equivalence classes are
ordered based on a reverse lexicographical ordering of their holes.