Support for templated impl declarations
The strategy that we use for now to support template instantiation is to
check the impl declaration as if it were a generic, but to defer all
checking of the impl definition until we see a use in which all template
parameters have arguments. At that point, we clone the impl definition
and type-check the whole thing, with constant values set on the template
parameters corresponding to the given arguments.

No caching of template instantiations is performed yet; each time we
form a reference to a template instantiation, we instantiate it afresh.