Bring constraints into scope when a generic calls a template.
When a template has an argument that involves a generic parameter, we're
supposed to delay instantiation until we know the concrete value, but
explorer is not set up to do that yet, so for now we instead instantiate
the template with the symbolic argument. When that happens, bring the
constraints on the generic parameter into scope so they can be used
inside the template instantiation.

This requires adding a new search over a value for the generic
parameters that appear within it; a `VisitNestedValues` visitor is added
to visit all the `Value`s nested with a value, and also convert an
existing place where we were doing the same thing in a way that was
incorrect (but harmlessly incorrect for now) to use it.