Parsing support for `if` expressions.
We model `if a then b else` as a prefix operator for parsing precedence
purposes. The rule that a statement starting with `if` is never an `if`
expression is handled implicitly because the statement parser never
invokes the expression parser for a statement starting with `if`.

This exposed a bug in our diagnosis of the whitespace rule for prefix
operators, which was incorrectly being applied to non-symbolic operators
in some cases, and was producing a bogus second diagnostic in some
cases, which is also fixed here.