Add a rich command-line argument parsing library.
This library is designed around supporting the kinds of use cases we
expect in the toolchain and other Carbon tools. It supports subcommands,
options, and prints help.

There is still a decent chunk of work to be done to finish polishing
this, but it should give us a solid starting point.

For details about the library and a brief roadmap, see the main comment
in `argparse.h` which provides a comprehensive overview of the library
and a roadmap of the remaining work.

Porting the driver to this went fairly well, but did require some
changes. Most visible are the changes to our recently added (yay!)
codegen layer. I've tried to make the APIs there a bit easier to
integrate, and now that we have this layer fleshed out the command line
interface for codegen somewhat. Happy to shuffle this arround.

Sadly, the codegen changes also showcase some limitations with our
current test infrastructure. The previous structure of the code gave
a few ways to workaround this, but those don't work any more. I'm not
going to great lengths to get everything tested here. Some of the
driver's behavior may need to be tested more thoroughly with follow-ups
to provide the necessary infrastructure like the ability to copy test
inputs into the temporary test directory.

One open question is whether `Args` is the best top-level name, or it
should switch to `CLI` or something else. Suggestions welcome!