Introduce a new Conduct team
The Carbon Project relied on the three leads to handle conduct concerns
initially as it bootstrapped its community team and expertise. We now
have an active and effective community lead and team of moderators. Our
community lead has worked to train and ramp up a new and independent
conduct team.

This proposal both provides an overview of the process and hands off
conduct handling to the new team! Going forward, we expect routine
updates to the conduct team to happen without full proposals as they
allow trained folks to rotate in and out of this difficult but essential
role on the project.

Co-authored-by: Céline Dedaj <CelineausBerlin@users.noreply.github.com>